Add Autofact, Aspect Validation
Autofac, Aspect foundations builder Validation Aspect class tested

+++ Entities/DTO/ProductDto.cs
@@ -15,7 +15,7 @@ namespace SmartPro.Entities.DTO
         public string? BrandName { get; set; }
         public string? ProductName { get; set; }
         public string? ProductDescription { get; set; }
-        public short Stock { get; set; }
+        public int Stock { get; set; }
         public decimal Price { get; set; }
     }
 }
